backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
langchain-openai==0.3.28
openai==1.97.0
pydantic==2.11.7
python-multipart==0.0.6
requests==2.31.0 